chore(build): Enable and configure Compose Compiler reports
This commit enables and configures Compose Compiler metrics and reports to help analyze and debug composition performance. The configuration logic is centralized into a new `configureComposeCompiler` extension function.

### Key Changes:

-   **Enable Reports:** Set `enableComposeCompilerMetrics` and `enableComposeCompilerReports` to `true` in `gradle.properties`.
-   **Centralized Configuration:**
    -   A new file, `ComposeCompilerConfiguration.kt`, is introduced, containing an internal `configureComposeCompiler` extension function.
    -   This function reads the new properties and sets the output directories for metrics and reports relative to each project's build folder.
    -   It also configures the stability configuration file (`compose_compiler_config.conf`).
-   **Build Logic Refactoring:**
    -   The `AndroidCompose.kt` and `KmpLibraryComposeConventionPlugin.kt` files are updated to call the new `configureComposeCompiler` function, removing duplicated configuration code.
-   **Dependency Cleanup:**
    -   Removed `toolchainVendor=JETBRAINS` from `gradle/gradle-daemon-jvm.properties` as it is redundant.
    -   Removed an unused `android.lint` plugin from `build-logic/convention/build.gradle.kts`.
