+++ gradle.properties
@@ -63,4 +63,8 @@ dependency.analysis.print.build.health=true
 ksp.incremental=true
 ksp.incremental.classpath=true
 
+# Compose Compiler Reports
+enableComposeCompilerMetrics=true
+enableComposeCompilerReports=true
+
 android.newDsl=false
